+++ lib/in_sync_interface.dart
@@ -1,2 +1,3 @@
+export 'src/sync_data_info.dart';
 export 'src/sync_repository.dart';
 export 'src/sync_state_info.dart';