+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [1.8.0]
 
-* More detailed synchronization status and upload progress.
+* More detailed synchronization status and send changes progress.
+   * Added enum `SyncDataState` with detailed data synchronization status.
+   * Added `SyncStateInfo.dataInfo` getter.
 
 ## [1.7.0]